virt: smbios : move fw_build_smbios() to smbios/smbios.c
fw_build_smbios() move to hw/smbios/smbios.c file for
CONFIG_SMBIOS and add stub function for disable smbios.
